Move tracing methods and constants from GenerateJLIClassesHelper to MethodHandleStatics
